# Include directories in ls recursive results.