--- a/tiledb/api/c_api/vfs/vfs_api.cc
+++ b/tiledb/api/c_api/vfs/vfs_api.cc
@@ -314,6 +314,23 @@ capi_return_t tiledb_vfs_ls_recursive(
   return TILEDB_OK;
 }
 
+capi_return_t tiledb_vfs_ls_recursive_v2(
+    tiledb_vfs_t* vfs,
+    const char* path,
+    tiledb_ls_callback_v2_t callback,
+    void* data) {
+  ensure_vfs_is_valid(vfs);
+  if (path == nullptr) {
+    throw CAPIStatusException("Invalid TileDB object: VFS passed a null path.");
+  } else if (callback == nullptr) {
+    throw CAPIStatusException(
+        "Invalid TileDB object: Callback function is null.");
+  }
+  ensure_output_pointer_is_valid(data);
+  vfs->ls_recursive_v2(tiledb::sm::URI(path), callback, data);
+  return TILEDB_OK;
+}
+
 }  // namespace tiledb::api
 
 using tiledb::api::api_entry_context;
@@ -556,5 +573,16 @@ CAPI_INTERFACE(
     tiledb_ls_callback_t callback,
     void* data) {
   return api_entry_context<tiledb::api::tiledb_vfs_ls_recursive>(
+      ctx, vfs, path, callback, data);
+}
+
+CAPI_INTERFACE(
+    vfs_ls_recursive_v2,
+    tiledb_ctx_t* ctx,
+    tiledb_vfs_t* vfs,
+    const char* path,
+    tiledb_ls_callback_v2_t callback,
+    void* data) {
+  return api_entry_context<tiledb::api::tiledb_vfs_ls_recursive_v2>(
       ctx, vfs, path, callback, data);
 }