This updates ls_recursive to include directories (prefixes) for object storage backends. 

Some performance testing before and after this change built in release mode, I don't see a major performance impact using these test buckets.

```
# Before
Fetched 464846 objects in 60343.9ms from S3 URI: s3://tiledb-shaun
Fetched 6868 objects in 1125.57ms from S3 URI: s3://tiledb-deep-archive
Fetched 998918 objects in 145453ms from S3 URI: s3://genomic-datasets/vcf/giab/datasets/structural-variants/

# After
Fetched 474641 objects in 52456.9ms from S3 URI: s3://tiledb-shaun
Fetched 11929 objects in 1180.48ms from S3 URI: s3://tiledb-deep-archive
Fetched 1068657 objects in 137803ms from S3 URI: s3://genomic-datasets/vcf/giab/datasets/structural-variants/
```

---
TYPE: FEATURE
DESC: Include directories in ls recursive results.