@teo-tsirpanis 2025-07-31
We should also open a story to audit all usages of `ls_recursive` in downstream components, and make sure they are prepared for this change.